Do not rely on fallback_version in setuptools_scm
As it is buggy. Just catch the exception.

Also lazy import setuptools_scm. It imports pkg_resources which is very
slow, though then again we're importing setuptools anyway so we're
probably already paying that cost. No matter, this is my standard
`__version__.py` at the moment so I won't deviate from it.