Add entry points. Clean build dir
If build dir is not cleaned, then building for multiple different
Pythons will include extensions compiled for one Python version in the
wheels for another version, which conda does not like one bit
(among other reasons not to do it):

https://github.com/pypa/wheel/issues/147

Add `--no-deps` to pip command. It doesn't seem to be necessary, as I
don't see dependencies being installed at build time. However it is
recommended so I'm doing it just in case.