Add a command to distutils
Obviating the need for packages to explicitly import and set up the
`dist_conda` command in their `setup.py`

This means we're no longer python and OS-independent, since the location
of the distutils module varies by OS and Python version.

So long as we can get GitHub building our packages, that's fine.
Otherwise might need to revisit this decision.