Workaround conda seting the wrong PATH on Windows Git Bash
Seems to be this issue:
https://github.com/conda/conda/issues/8693

Conda adds Miniconda\Scripts\condabin (which doesn't exist) to PATH
instead of Miniconda3\condabin. This is fine if you're in the shell
where 'conda' is some kind of shell alias or function so it can still be
run. But it's a problem if you want to call conda programmatically from
a non-shell process.