Use cmd.exe for conda on Windows.
Conda with git bash has too many bugs, such as
https://github.com/conda/conda/issues/8693 which means condabin is
sometimes not in the PATH, and when you run a program it does incorrect
conversion of PATH, turning some colons (e.g. in C://) into
semicolons, splitting the paths and making them useless. Much breakage
results from this.

Also, use shell=True in subprocess calls to conda. I have no idea why
this is needed, but without it, conda commands mysteriously fail even
though they appear to succeed. Perhaps they are operating on the wrong
environment.

Maybe this path nonsense with git bash, or the shell=True requirement,
is the cause of not getting compilation of extensions working on Windows
when running on GitHub actions? I hope so.