Have setuptools-conda build use name differences.
Allow it to use all the options it uses from the command line

Have it prioritise dist_conda specific options in setup.cfg over
pyproject.toml, and remove the option to specify custom channels in
pyproject.toml. We can put things there if we make a wholesale move to
that file, but we're a setuptools tool so we won't do that until
setuptools itself does.