extras_require support
Added support for specifying project paths with the extras_require modifier available when doing a pip install

For example, you should now be able to runs something like:
`setuptools_conda install-requirements labscript-suite[docs]`
which should gather and include the extra requirements under the "docs" category.

As it is actually project paths being specified in the command, the new code distinguishes between paths with square brackets in and the extras modifier.