Don't specify shell as global default
If we add any steps in future that aren't using conda python, then on
macos these will cause `python` to point to system python (python 2!)
rather than a python installed with the "install-python" action.

We're not likely to do that in this project, but doing it anyway for
consistency with workflows of other related projects.

Also, setup-miniconda is save to be v2 for the `upload` job since that
doesn't run on Windows.