Fix incorrect sdist filename
This broke building packages with dashes in their names from source,
including this one.

As of a recent release of `setuptools`, sdist filenames now are in the
format `package_name-<version>.tar.gz` instead of
`package-name-<version>.tar.gz` (i.e. underscores in the package name
instead of dashes). This is the case even if the package name itself
has dashes in it.

https://github.com/pypa/setuptools/issues/4214

https://github.com/pypa/setuptools/issues/4167

This change is seemingly deliberate in order to conform to a standard
for filenames specifically even though the standard for project names
is increasingly to use dashes.